# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Clinic Service Assistant API
  version: 0.1.0
  description: |
    Чат-бот подбора услуг клиники. Порядок обработки сообщения пациента:
    1) детерминированный триаж red (triage_rules.json) — без LLM, по ВСЕМ сообщениям пациента;
       при флаге беременности добавляются правила red_if_pregnant;
    2) LLM-извлечение (language, intent, specialty_id, urgency, ...);
    3) уточнения (макс. 2 на диалог);
    4) подбор услуг и врачей из каталога;
    5) ответ LLM только по данным каталога;
    6) передача оператору (обращение с summary).
servers:
  - url: http://localhost:8088
security:
  - bearer: []
paths:
  /api/auth/patient:
    post:
      summary: Анонимная сессия пациента
      security: []
      responses:
        "200": {description: Токен, content: {application/json: {schema: {$ref: "#/components/schemas/Token"}}}}
  /api/auth/login:
    post:
      summary: Вход оператора
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties: {username: {type: string, example: operator1}, password: {type: string, example: operator1}}
      responses:
        "200": {description: Токен, content: {application/json: {schema: {$ref: "#/components/schemas/Token"}}}}
        "401": {description: Неверный логин/пароль}
  /api/chat:
    post:
      summary: Сообщение пациента (роль patient)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                dialog_id: {type: string, format: uuid, description: Пусто — создать новый диалог}
                message: {type: string, maxLength: 4000}
      responses:
        "200": {description: Ответ бота, content: {application/json: {schema: {$ref: "#/components/schemas/ChatResponse"}}}}
        "400": {description: Пустое сообщение}
        "401": {description: Нет токена}
        "404": {description: Диалог не найден или чужой}
  /api/chat/operator:
    post:
      summary: Кнопки «Связаться с оператором» (contact_operator) и «Срочно связаться с оператором» (urgent_operator)
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [dialog_id], properties: {dialog_id: {type: string, format: uuid}}}
      responses:
        "200":
          description: Обращение создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket_id: {type: integer}
                  status: {type: string}
                  urgent: {type: boolean, description: true — диалог red (кнопка «Срочно связаться с оператором»)}
                  reply: {$ref: "#/components/schemas/Message"}
  /api/chat/attachments:
    post:
      summary: Пациент прикрепляет фото/PDF анализа или заключения (бот файл не читает — его смотрит врач/оператор)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary, description: "JPG, PNG, WEBP или PDF до 10 МБ; тип проверяется по содержимому"}
                dialog_id: {type: string, format: uuid, description: Пусто — новый диалог}
      responses:
        "200":
          description: Файл сохранён
          content:
            application/json:
              schema:
                type: object
                properties:
                  dialog_id: {type: string, format: uuid}
                  status: {type: string}
                  message: {$ref: "#/components/schemas/Message"}
                  reply: {$ref: "#/components/schemas/Message"}
        "413": {description: Больше 10 МБ}
        "415": {description: Неподдерживаемый тип}
  /api/attachments/{id}:
    get:
      summary: Файл вложения. Пациент — только свои, оператор — только из переданных диалогов
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Содержимое файла с его Content-Type}
        "404": {description: Нет файла или нет доступа}
  /api/dialogs/{id}:
    get:
      summary: Диалог с сообщениями. Пациент — только свой; оператор — только переданные (есть обращение).
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: Диалог
          content:
            application/json:
              schema:
                type: object
                properties:
                  dialog: {$ref: "#/components/schemas/Dialog"}
                  messages: {type: array, items: {$ref: "#/components/schemas/Message"}}
                  tickets: {type: array, items: {$ref: "#/components/schemas/Ticket"}, description: Только для оператора}
        "404": {description: Не найден}
  /api/operator/queue:
    get:
      summary: Очередь обращений (роль operator), сортировка red → yellow → green, затем по времени
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [open, closed, all], default: open}}
      responses:
        "200": {description: Очередь, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/Ticket"}}}}}
        "403": {description: Не оператор}
  /api/operator/reply:
    post:
      summary: Ответ оператора в диалог; close=true закрывает обращение и возвращает диалог боту
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dialog_id]
              properties:
                dialog_id: {type: string, format: uuid}
                message: {type: string}
                close: {type: boolean, default: false}
      responses:
        "200": {description: OK}
  /api/catalog:
    get:
      summary: Каталог (специальности, услуги, врачи со слотами)
      security: []
      responses:
        "200": {description: Каталог, content: {application/json: {schema: {$ref: "#/components/schemas/Catalog"}}}}
  /api/appointments:
    post:
      summary: Запись к врачу (кнопка «Записаться», роль patient)
      description: |
        Слот должен быть среди свободных слотов врача в каталоге, услуга — той же специальности, что и врач.
        После записи слот исчезает из /api/catalog. Если передан dialog_id (свой диалог), в него добавляется
        сообщение бота «Вы записаны: <услуга>, <врач>, <дата время>.» на языке диалога (ru/kk/en).
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/AppointmentCreate"}}}
      responses:
        "201":
          description: Запись создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointment: {$ref: "#/components/schemas/Appointment"}
                  reply: {$ref: "#/components/schemas/Message"}
        "400": {description: "Неизвестный врач/услуга, услуга не той специальности, слота нет в расписании, имя не 2–100 символов, телефон не казахстанский", content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "403": {description: Не пациент}
        "404": {description: dialog_id не найден или чужой}
        "409": {description: Слот уже занят, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    get:
      summary: "Записи: пациент — только свои, оператор — все. Сначала новые."
      responses:
        "200": {description: Список, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/Appointment"}}}}}
  /api/appointments/{id}/cancel:
    post:
      summary: "Отмена записи (пациент — своей, оператор — любой). Слот возвращается врачу."
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        "200": {description: Запись отменена, content: {application/json: {schema: {$ref: "#/components/schemas/Appointment"}}}}
        "404": {description: Запись не найдена или чужая}
        "409": {description: Запись уже отменена}
  /api/integration/catalog:
    put:
      tags: [integration]
      summary: "МИС: полная замена каталога (атомарно)"
      description: |
        Тело — каталог в формате catalog.json. Проверяется: уникальность id, ссылки услуг и врачей на существующие
        специальности, формат слотов YYYY-MM-DDTHH:MM (дубликаты слотов убираются, слоты сортируются).
        Уже занятые записями слоты в каталог не возвращаются.
      security: [{apiKey: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/Catalog"}}}
      responses:
        "200":
          description: Каталог заменён
          content: {application/json: {schema: {type: object, properties: {specialties: {type: integer}, services: {type: integer}, doctors: {type: integer}}}}}
        "400": {description: Неверный каталог (причина в error), content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Нет или неверный X-API-Key}
        "503": {description: Интеграция выключена (INTEGRATION_API_KEY не задан)}
  /api/integration/doctors/{id}/slots:
    put:
      tags: [integration]
      summary: "МИС: замена свободных слотов одного врача"
      description: Слоты, уже занятые записями (status=booked), отфильтровываются.
      security: [{apiKey: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string}, example: d1}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slots]
              properties: {slots: {type: array, items: {type: string, example: "2026-10-01T09:00"}}}
      responses:
        "200":
          description: Итоговые свободные слоты
          content: {application/json: {schema: {type: object, properties: {doctor_id: {type: string}, slots: {type: array, items: {type: string}}}}}}
        "400": {description: Неверный формат слотов}
        "401": {description: Нет или неверный X-API-Key}
        "404": {description: Врач не найден}
        "503": {description: Интеграция выключена}
  /api/integration/appointments:
    get:
      tags: [integration]
      summary: "МИС: выгрузка записей для импорта"
      description: |
        Записи, созданные или изменённые (например, отменённые) начиная с since (включительно), по возрастанию
        updated_at. Для инкрементальной синхронизации передавайте наибольший полученный updated_at; дубликаты
        устраняйте по id. Без since — все записи (до 5000).
      security: [{apiKey: []}]
      parameters: [{name: since, in: query, required: false, schema: {type: string, format: date-time}, example: "2026-10-01T00:00:00Z"}]
      responses:
        "200": {description: Записи, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/Appointment"}}}}}
        "400": {description: since не в формате RFC3339}
        "401": {description: Нет или неверный X-API-Key}
        "503": {description: Интеграция выключена}
  /api/health:
    get:
      security: []
      summary: Проверка живости
      responses:
        "200": {description: OK}
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
    apiKey: {type: apiKey, in: header, name: X-API-Key, description: "Ключ МИС для /api/integration/* (env INTEGRATION_API_KEY)"}
  schemas:
    Urgency: {type: string, enum: [green, yellow, red]}
    Error:
      type: object
      properties: {error: {type: string}}
    Specialty:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        description: {type: string}
        treats: {type: array, items: {type: string}}
        i18n: {$ref: "#/components/schemas/I18n"}
    Catalog:
      type: object
      required: [specialties, services, doctors]
      properties:
        specialties: {type: array, items: {$ref: "#/components/schemas/Specialty"}}
        services: {type: array, items: {$ref: "#/components/schemas/Service"}}
        doctors: {type: array, items: {$ref: "#/components/schemas/Doctor"}, description: slots — только свободные}
    AppointmentCreate:
      type: object
      required: [doctor_id, service_id, slot, patient_name, phone]
      properties:
        doctor_id: {type: string, example: d1}
        service_id: {type: string, example: ther-consult}
        slot: {type: string, example: "2026-10-01T09:00", description: один из свободных слотов врача}
        dialog_id: {type: string, format: uuid, description: необязательно; свой диалог — в него придёт подтверждение}
        patient_name: {type: string, minLength: 2, maxLength: 100}
        phone: {type: string, example: "+7 701 123-45-67", description: "+7 или 8 и 10 цифр; пробелы, дефисы, скобки допускаются"}
    Appointment:
      type: object
      properties:
        id: {type: string, format: uuid}
        dialog_id: {type: string, format: uuid, nullable: true}
        patient_id: {type: string}
        doctor_id: {type: string}
        service_id: {type: string}
        slot: {type: string, example: "2026-10-01T09:00"}
        patient_name: {type: string}
        phone: {type: string}
        status: {type: string, enum: [booked, cancelled]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Token:
      type: object
      properties: {token: {type: string}, user_id: {type: string}, role: {type: string, enum: [patient, operator]}}
    Service:
      type: object
      properties:
        id: {type: string}
        specialty_id: {type: string}
        name: {type: string, description: на русском}
        price: {type: integer, description: тенге}
        description: {type: string, description: на русском}
        i18n: {$ref: "#/components/schemas/I18n"}
    I18n:
      type: object
      description: "Переводы: {kk: {name, description}, en: {name, description}}. Фронтенд показывает поле на языке интерфейса, иначе базовое (ru)."
      additionalProperties:
        type: object
        properties: {name: {type: string}, description: {type: string}}
    Doctor:
      type: object
      properties: {id: {type: string}, name: {type: string}, specialty_id: {type: string}, slots: {type: array, items: {type: string, example: "2026-10-01T09:00"}}}
    Message:
      type: object
      properties:
        id: {type: integer}
        role: {type: string, enum: [patient, bot, operator]}
        author: {type: string}
        content: {type: string}
        data:
          type: object
          description: Для сообщений бота
          properties:
            urgency: {$ref: "#/components/schemas/Urgency"}
            actions:
              type: array
              description: "call_103 — кнопка tel:103; contact_operator — «Связаться с оператором»; urgent_operator — «Срочно связаться с оператором» (red у беременной). Обе операторские кнопки вызывают POST /api/chat/operator."
              items: {type: string, enum: [call_103, contact_operator, urgent_operator]}
            services: {type: array, items: {$ref: "#/components/schemas/Service"}}
            doctors: {type: array, items: {$ref: "#/components/schemas/Doctor"}}
            attachments:
              type: array
              items:
                type: object
                properties: {id: {type: string, format: uuid}, name: {type: string}, content_type: {type: string}, size: {type: integer}}
        created_at: {type: string, format: date-time}
    Dialog:
      type: object
      properties:
        id: {type: string, format: uuid}
        patient_id: {type: string}
        status: {type: string, enum: [bot, operator]}
        language: {type: string, enum: [ru, kk, en]}
        urgency: {$ref: "#/components/schemas/Urgency"}
        urgency_reason: {type: string}
        clarifications: {type: integer, maximum: 2}
        summary: {type: string}
        pregnant: {type: boolean, description: "Флаг беременности: ставится кодом по фразам или LLM, никогда не снимается"}
        gestation_weeks: {type: integer, nullable: true}
        pregnancy_asked: {type: boolean, description: Вопрос о беременности уже задан}
        risk_factors: {type: array, items: {type: string, enum: [hypertension, diabetes, anemia, heart_disease, kidney_disease, endocrine, multiple_pregnancy, preterm_risk, previous_complications]}, description: "Акушерские факторы риска, названные пациенткой; только добавляются. Беременность + фактор риска + жалоба → не ниже yellow"}
        assessed: {type: boolean, description: "Срочность оценена и показана пациенту. До этого пациент получает urgency: \"\" (бот сначала задаёт уточняющий вопрос); red оценивается сразу"}
        created_at: {type: string, format: date-time}
    Ticket:
      type: object
      properties:
        id: {type: integer}
        dialog_id: {type: string, format: uuid}
        reason: {type: string, description: Почему передано оператору}
        summary: {type: string, description: Что хотел пациент и что выяснено}
        status: {type: string, enum: [open, closed]}
        urgency: {$ref: "#/components/schemas/Urgency"}
        language: {type: string}
        pregnant: {type: boolean}
        gestation_weeks: {type: integer, nullable: true}
        risk_factors: {type: array, items: {type: string}}
        last_message: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    ChatResponse:
      type: object
      properties:
        dialog_id: {type: string, format: uuid}
        status: {type: string, enum: [bot, operator]}
        urgency: {type: string, enum: [green, yellow, red, ""], description: "Пусто, пока срочность не оценена (первый ответ бота на жалобу — уточняющий вопрос)"}
        urgency_reason: {type: string}
        language: {type: string, enum: [ru, kk, en]}
        reply: {allOf: [{$ref: "#/components/schemas/Message"}], nullable: true, description: null — диалог у оператора, бот молчит}
        actions: {type: array, items: {type: string, enum: [call_103, contact_operator, urgent_operator]}}
        services: {type: array, items: {$ref: "#/components/schemas/Service"}}
        doctors: {type: array, items: {$ref: "#/components/schemas/Doctor"}}
        ticket_id: {type: integer, nullable: true}